Major refactoring of the code extracting cumulative distributions.
The main algorithm is the same.  Some functions were simplified, some
type safety was added (mostly as a mean to help me retro-engeneer),
and Bioconductor types have been preferred to ad-hoc construct when
no obvious penalty was seen on the performance.

Note that the code for working on consensus clusters is left broken
at the moment, waiting for the work on cluster aggregation to start.