Repair paraclu support for CAGEset objects
Closes (#8)

This commit bypasses the regression caused by introducing
Bioconductor classes in the clustering functions for
CAGEset objects.  CAGEexp object will need to wait for
a rework of the code in order to have paraclu support.

Also, it is possible to run paraclu externally, and to
feed the cluster information in the CustomConsensusClusters
function.