fitPowerLaw: S4 function accepting various CAGEr objects.
The goal is to easily record powerLaw fitting parameters in a CAGEexp object.

Also change .fit.power.law.to.reverse.cumulative to use Rle instead of
data.table to compute the reverse cumulative distributions.

Next: make the normalisation and plotting functions reuse the fitted values
if they already are found in the metadata.